pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "ideascout"
version = "0.1.0"
description = "Local-first idea scouting from Reddit, HN, IndieHackers, and other sources"
requires-python = ">=3.13"
dependencies = [
    "feedparser>=6.0.10",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[project.scripts]
ideascout = "ideascout.cli:main"

[tool.setuptools.packages.find]
include = ["ideascout*"]
exclude = ["tests*", "data*", "dashboard*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
